+++ lib/instream/data/read.ex
@@ -7,11 +7,16 @@ defmodule Instream.Data.Read do
 
   @doc """
   Creates a reading query object from a raw query string.
+
+  Valid options for read queries are:
+
+  - `precision`: see `Instream.Encoder.Precision` for available values
   """
-  @spec query(String.t) :: Query.t
-  def query(query_str) do
+  @spec query(String.t, Keyword.t) :: Query.t
+  def query(query_str, opts \\ []) do
     %Query{
       payload: query_str,
+      opts:    opts,
       type:    :read
     }
   end