+++ lib/instream/query/read.ex
@@ -8,13 +8,14 @@ defmodule Instream.Query.Read do
   alias Instream.Query.Headers
   alias Instream.Query.URL
 
-  def execute(%Query{ payload: payload }, opts, conn) do
+  def execute(%Query{} = query, opts, conn) do
     headers = conn |> Headers.assemble()
     url     =
          conn
       |> URL.query()
       |> URL.append_database(opts[:database])
-      |> URL.append_query(payload)
+      |> URL.append_precision(query.opts[:precision])
+      |> URL.append_query(query.payload)
 
     { :ok, status, headers, client } = :hackney.get(url, headers)
     { :ok, response }                = :hackney.body(client)