+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enhancements
   - Convenience module for "SHOW DIAGNOSTICS" queries
   - Convenience module for "SHOW STATS" queries
+  - Precision (= epoch) can be passed to read queries
 
 - Bug fixes
   - Pool configuration (size, overflow) is now properly taken from configuration